+++ projects/iad-demo/src/app/faker/responses/cms/api/iad/presentation.response.ts
@@ -146,6 +146,7 @@ export class PresentationResponse extends AbstractResponse {
                     {
                         'field': 'actions',
                         'header': 'presentation.projection.main.actions',
+                        'sorting': false,
                         'translate': true,
                         'visible': true,
                         'formatter': 'actions',