+++ projects/iad-interface-admin/form/src/lib/dynamic-form/components/iad-form-input.component.ts
@@ -33,13 +33,13 @@ col-form-label">{{config.translate ? (config.label | translate) : config.label}}
                         <iad-tooltip-notifier *ngIf="!(config.readonly || config.disabled) && isInvalid"
                                               caption="Ошибка!" [text]="error" [activated]="true">
                             <ng-template>
-                                <fa-icon [icon]="'exclamation-circle'" [size]="'1x'" [ngStyle]="{color: 'red'}"></fa-icon>
+                                <iad-icon-outlet icon="fas exclamation-circle" [size]="'1x'" [cssStyle]="{color: 'red'}"></iad-icon-outlet>
                             </ng-template>
                         </iad-tooltip-notifier>
                         <iad-tooltip-notifier *ngIf="!(config.readonly || config.disabled) && !isInvalid"
                                               caption="Поле валидно!" text="Валидация прошла успешно" [activated]="true">
                             <ng-template>
-                                <fa-icon [icon]="'check-circle'" [size]="'1x'" [ngStyle]="{color: 'green'}"></fa-icon>
+                                <iad-icon-outlet icon="fas check-circle" [size]="'1x'" [cssStyle]="{color: 'green'}"></iad-icon-outlet>
                             </ng-template>
                         </iad-tooltip-notifier>
                     </div>