+++ projects/iad-interface-admin/form/src/lib/dynamic-form/inputs/dropdown-multi-input.model.ts
@@ -1,0 +1,11 @@
+import { DropdownInputModel } from './dropdown-input.model';
+
+/**
+ * Note! Prefer DropdownInputModel instead
+ */
+export class DropdownMultiInputModel extends DropdownInputModel {
+    constructor(options: {} = {}) {
+        super(options);
+        this.multiple = true;
+    }
+}